Kan rita nu och det command för redo och undo funktion är påbörjad. Main initializerar allt så att alla nödvändiga element fungerar från start.

+++ src/Model/Circle.java
@@ -2,7 +2,7 @@ package Model;
 
 public class Circle extends Shape{
 
-    public Circle(int x, int y, int width, int height, boolean isFilled){
+    public Circle(double x, double y, int width, int height, boolean isFilled){
         super(x,y,width,height,isFilled);
     }
 }
