+++ src/Model/Square.java
@@ -2,7 +2,7 @@ package Model;
 
 public class Square extends Shape{
 
-    public Square(int x, int y, int width, int height, boolean isFilled){
+    public Square(double x, double y, int width, int height, boolean isFilled){
         super(x,y,width,height,isFilled);
     }
 }
